Move celo event definitions into dedicated module.
